+++ composeApp/src/androidMain/kotlin/proj/memorchess/axl/MainActivity.kt
@@ -5,6 +5,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
+import proj.memorchess.axl.ui.KoinStarterApp
 
 lateinit var MAIN_ACTIVITY: MainActivity
