+++ composeApp/src/iosMain/kotlin/proj/memorchess/axl/MainViewController.kt
@@ -1,5 +1,6 @@
 package proj.memorchess.axl
 
 import androidx.compose.ui.window.ComposeUIViewController
+import proj.memorchess.axl.ui.KoinStarterApp
 
 fun MainViewController() = ComposeUIViewController { KoinStarterApp() }
