+++ composeApp/src/jvmMain/kotlin/proj/memorchess/axl/main.kt
@@ -2,6 +2,7 @@ package proj.memorchess.axl
 
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
+import proj.memorchess.axl.ui.KoinStarterApp
 
 fun main() = application {
   Window(onCloseRequest = ::exitApplication, alwaysOnTop = true, title = "Memor Chess") {
