+++ composeApp/src/wasmJsMain/kotlin/proj/memorchess/axl/main.kt
@@ -15,6 +15,7 @@ import kotlinx.browser.document
 import org.koin.compose.KoinApplication
 import org.koin.compose.koinInject
 import proj.memorchess.axl.core.data.online.auth.AuthManager
+import proj.memorchess.axl.ui.App
 import proj.memorchess.axl.ui.components.buttons.SignInButton
 
 @OptIn(ExperimentalComposeUiApi::class)
